Add simple file exporter for jar exec-mode
Further execution modes can be supported in the future by adding corresponding
implementation of the generic function `plantuml-export-string-to-buffer`.
Export currently fails if the target file name is already present, but this
should be customizable in the future.

The export mechanism could be used to also implement the preview functionality.
However, the export mechanism is current strictly synchronous and must be
changed to work asynchronously first.