Conduct diagram conversion in original buffer while exporting
The reason for this is that `plantuml-output-type` is local to the current
buffer, and calling `plantuml-export-string-to-buffer` in a temporary buffer may
see another value for this variable as originally set in the corresponding
plantuml buffer.  If, for example, the value is "svg" in the plantuml buffer,
but has a global value of "png", then using `with-temp-buffer` would result in a
png file with ending .svg.  This is undesired behavior, and can be fixed by
keeping the current buffer until it's written to a file.