Do not automatically render image on export
`write-file` may call `set-auto-mode` (e.g., when `lsp-mode` is active), which
in turn will call `image-mode` on the export buffer and will render the image in
Emacs itself.  When the image is large, this may cause a considerable amount of
delay before exporting is done, even if the image is not meant to be displayed
in Emacs itself (which is what `plantuml-preview` is for, actually).  Let's try
to disable that behavior.